requirements-fastapi.txt
# FastAPI and related packages
fastapi
uvicorn
python-multipart

# Asynchronous HTTP client
aiohttp
aiofiles

# Environment variables management
python-dotenv

# Asyncio utilities
nest_asyncio

# LangChain and related packages
langchain==1.0.0
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-community==0.4
langchain-tavily==0.2.12
# LangGraph and related packages
langgraph==1.0.0
langgraph-checkpoint==2.1.0
langgraph-checkpoint-cosmosdb==0.2.5

# Streamlit for web apps
streamlit==1.52.1
streamlit-cookies-controller==0.0.4

# Azure services
azure-storage-blob==12.27.1
azure-identity==1.25.1
azure-cosmos==4.14.3

# Image processing
Pillow
opencv-python

# PDF and document processing
PyMuPDF==1.26.4
fitz==0.0.1.dev2
pdf2image
pdfplumber
markitdown[all]
PyPDF2

# Miscellaneous utilities
einops
pytz
qdrant_client
bcrypt
pandas
numpy
requests
typing-extensions
pydantic
filetype